requirements.txt
aiohttp
opencv-python
mediapipe
numpy
speedtest-cli